plop-pack-internal: Rename helper function
`title` is too generic and is likely to clash with the user's variables.
Let's use `mdTitle` instead as this pertains to markdown specifically.
