+++ lib/create/skel/project/README.md.hbs
@@ -1,4 +1,4 @@
-{{#title}}{{{fullName}}}{{/title}}
+{{#mdTitle}}{{{fullName}}}{{/mdTitle}}
 
 {{{description}}}
