+++ lib/create/skel/prototype/README.md.hbs
@@ -1,4 +1,4 @@
-{{#title}}{{{titleCase name}}}{{/title}}
+{{#mdTitle}}{{{titleCase name}}}{{/mdTitle}}
 
 {{{description}}}
