+++ lib/plop-pack/skel/app/README.md.hbs
@@ -1,4 +1,4 @@
-{{#title}}{{{titleCase (pkg 'name')}}} - {{{titleCase name}}}{{/title}}
+{{#mdTitle}}{{{titleCase (pkg 'name')}}} - {{{titleCase name}}}{{/mdTitle}}
 
 {{{description}}}
