+++ lib/plop-pack/src/helpers/md-title.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const title = function(options) {
+const mdTitle = function(options) {
   const t = options.fn(this);
   return `${t}\n${t.replace(/./g, '=')}`;
 }
 
-module.exports = title;
+module.exports = mdTitle;
