+++ lib/plop-pack/src/index.js
@@ -12,7 +12,7 @@ const plopGeneratorComponent = require('./generators/component');
 const plopGeneratorLibrary = require('./generators/lib');
 
 const handlebarsHelperEq = require('./helpers/eq');
-const handlebarsHelperTitle = require('./helpers/title');
+const handlebarsHelperMdTitle = require('./helpers/md-title');
 
 const plopFunction = plop => {
   plop.setActionType('copy', plopActionCopy);
@@ -27,7 +27,7 @@ const plopFunction = plop => {
   plop.setGenerator('lib', plopGeneratorLibrary);
 
   plop.setHelper('eq', handlebarsHelperEq);
-  plop.setHelper('title', handlebarsHelperTitle);
+  plop.setHelper('mdTitle', handlebarsHelperMdTitle);
 
   plop.setDefaultInclude({ actionTypes: true, generators: true, helpers: true });
 };
